Algoryn and Ghar Rebel updates
Fixed an issue of Algoryn support and specialist support were forced to all take an upgrade of to their weapon if any wanted to. Rather than being able to take individual upgrades as the book says for the unit.
Also fixed issues of Rebel Ghar Command Battle Suits having issues with Plasma Claws, and the same issue as above with the Creeper making the 1-3 of them all having to take a weapon change if 1 did (again they can be upgraded individually).